--- src/models/items/index.ts
@@ -1,1 +1,0 @@
-export * from './items';
