--- src/models/items/items.ts
@@ -1,13 +1,0 @@
-
-export const basic_resources = [
-    'tVRP',
-    'spore',
-    'spice',
-    'metal',
-    'biomass',
-    'carbon'
-] 
-
-export async function geyUserItemBalance (itenId: number, userId: number): Promise<number> {
-    return 0
-}
