+++ src/models/telegram/index.ts
@@ -1,4 +1,4 @@
 export  * from "./duel";
 export  * from "./personal";
-export  * from "./store";
+export  * from "../store/store";
 export  * from "./subscribe";
